% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViScoreR.R
\name{NeighbourhoodComposition}
\alias{NeighbourhoodComposition}
\title{Compute neighbourhood composition}
\usage{
NeighbourhoodComposition(X, annot, pop, k = 1000, stepsize = 10, exclude = c())
}
\arguments{
\item{X}{numeric matrix: row-wise coordinate matrix of high-dimensional data or lower-dimensional embedding of it}

\item{annot}{string/factor vector: vector of labels per row of \code{X}}

\item{pop}{string: name of reference cell population}

\item{k}{integer: size of neighbourhood (default: 1000)}

\item{stepsize}{integer: increment in neighbourhood size per which to aggregate neighbourhood composition}

\item{exclude}{string vector: labelled populations to exclude (default: empty)}
}
\description{
Computes composition of near neighbourhood of a labelled population in terms of other populations in the data.
To that end, a separate neighbour search that excludes unassigned cells and cells from the reference population is conducted.
This means, for instance, that a neighbourhood of 'Monocytes' excluding 'Monocytes' themselves, as well as 'Debris/Doublets' is searched.
}
\details{
The computed neighbourhood composition is then relevant for a specific dataset, a population within it and either the original high-dimensional data or a specific embedding of it.
}
