% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViScoreR.R
\name{PlotEmbedding}
\alias{PlotEmbedding}
\title{Plot 2-d embedding of single-cell data}
\usage{
PlotEmbedding(
  X,
  annot,
  unassigned = c(),
  palette = .Palette(),
  pointsize = 0.5,
  no_legend = FALSE,
  font_family = NULL
)
}
\arguments{
\item{X}{numeric matrix: row-wise coordinate matrix of 2-dimensional embedding}

\item{annot}{string/factor vector: vector of labels per row of \code{X}}

\item{unassigned}{string (vector): labels in `annot` that belong to cells of unknown identity/debris/doublets (default: none)}

\item{palette}{string vector: palette of colours per population to use for plotting (default palette is provided)}

\item{pointsize}{numeric: size of each point in the plot (default: 0.5)}

\item{no_legend}{logical: disable displaying the legend mapping colours to populations? (default: FALSE)}

\item{font_family}{optional string: non-default font family to use for plotting}
}
\description{
Plots the embedding result of a dimension-reduction algorithm, with cell labels.
If any 'unassigned' labels are specified, the unassigned points are plotted in light grey in the background
}
