% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViScoreR.R
\name{PlotLD}
\alias{PlotLD}
\title{Plot likeness distributions}
\usage{
PlotLD(xnpe, pop, bin_width = 10, font_family = NULL)
}
\arguments{
\item{xnpe}{xNPE object, result of the \code{ViScoreR::xNPE} function call}

\item{pop}{string: name of population}

\item{bin_width}{integer: width of bins (like-neighbour count values from 0 to k+1) for aggregating the data (default: 10)}

\item{font_family}{optional string: font family to use for plotting}
}
\description{
Plots the likeness distributions (from xNPE) for a specific population in original data and an embedding.
}
