% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViScoreR.R
\name{PlotNC}
\alias{PlotNC}
\title{Plot neighbourhood composition}
\usage{
PlotNC(
  nc,
  stepsize = 10,
  exclude = c(),
  k = NULL,
  palette = .Palette(),
  method = NULL,
  no_legend = FALSE,
  font_family = NULL
)
}
\arguments{
\item{nc}{neighbourhood composition object, result of \code{ViScoreR::NeighbourhoodComposition} function call}

\item{stepsize}{integer: increment in neighbourhood size per which to aggregate neighbourhood composition}

\item{exclude}{string vector: labelled populations to exclude (default: empty)}

\item{k}{optional integer: size of neighbourhood cut-off}

\item{palette}{string vector: palette of colours per population to use for plotting (default palette is provided)}

\item{method}{optional string: name of the dimensionality reduction method or original dataset to include in title (default: none)}

\item{no_legend}{logical: disable displaying the legend mapping colours to populations? (default: FALSE)}

\item{font_family}{optional string: non-default font family to use for plotting}
}
\description{
Plots neighbourhood composition for a population of interest, as computed by function \code{NeighbourhoodComposition}.
A stacked area plot that is smoothed along the x-axis (neighbourhood size) is produced.
(The plot is not cumulative along the x-axis.)
}
