% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViScoreR.R
\name{PlotxNPE}
\alias{PlotxNPE}
\title{Plot xNPE per population}
\usage{
PlotxNPE(
  ...,
  methods = c("default"),
  exclude = c(),
  polar = FALSE,
  palette = .Palette(),
  font_family = NULL
)
}
\arguments{
\item{...}{one or more xNPE objects, results of \code{ViScoreR::xNPE} function calls}

\item{methods}{string vector: name of each xNPE result provided}

\item{exclude}{string vector: labelled populations to exclude (default: empty)}

\item{polar}{logical: use polar coordinates for plotting? (default: FALSE)}

\item{palette}{string vector: palette of colours per population to use for plotting (default palette is provided)}

\item{font_family}{optional string: non-default font family to use for plotting}
}
\description{
Plots one or more xNPE results for a single dataset.
xNPE values for each population are shown, and can be compared across multiple embedding methods.
}
