% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViScoreR.R
\name{xNPE}
\alias{xNPE}
\title{Compute xNPE (Extended Neighbourhood-Proportion-Error)}
\usage{
xNPE(hd, ld, annot, knn_hd = NULL, knn_ld = NULL, k = NULL, exclude = c())
}
\arguments{
\item{hd}{numeric matrix: row-wise coordinate matrix of high-dimensional data}

\item{ld}{numeric matrix: row-wise coordinate matrix of low-dimensional embedding of \code{hd}}

\item{annot}{string/factor vector: vector of labels per row of \code{hd}, \code{ld}}

\item{knn_hd}{optional integer matrix: k-nearest-neighbour matrix of \code{hd}, result of \code{ViScoreR::ComputekNN} function call}

\item{k}{integer: size of neighbourhood if \code{knn_hd} or \code{knn_ld} are unspecified}

\item{exclude}{string vector: labelled populations to exclude (default: empty)}
}
\description{
Computes xNPE (based in the Neighborhood Proportion Error) for dimensionality reduction (DR) result.
The xNPE gives a shape distortion score for each labelled population, computed in terms of how the proportion of alike versus not-alike cells changes in the reference population's neighbourhood as we go from the original data to the DR embedding.
}
\details{
It is advisable to pre-compute the k-nearest-neighbour graph for the original data as well as the embedding and provide them to this function.
(Use the \code{ViScoreR::ComputekNN} function for this.)
}
\references{
Konstorum A, Jekel N, Vidal E, Laubenbacher R (2018) Comparative Analysis of Linear and Nonlinear Dimension Reduction Techniques on Mass Cytometry Data, bioRxiv 273862; doi: https://doi.org/10.1101/273862
}
